fix: Claim failure in-case of less than min delegation
The claims weren't working for the withdrawals created on day-1 when the deposits on day-1 were less than the minimum delegation amount. This was because the initiateDelegation() wasn't called by bot in such cases because of the minimum delegation requirement.

Added a fix to short-circuit the bot for this edge-case.