fix: Testnet findings (#10)
Fixes MONDAY/2533020518/2920442054
Fixes MONDAY/2533020518/2922816596

This PR fixes the following things:
### Claim failure in-case of less than min delegation
The claims weren't working for the withdrawals created on day-1 when the deposits on day-1 were less than the minimum delegation amount. This was because the initiateDelegation() wasn't called by bot in such cases because of the minimum delegation requirement. Added a fix to short-circuit the bot for this edge-case.

### unbondingFinished failure
There was no receive ether function in the StakePool contract. Resulting in a failure when the UndelegationHolder was trying to send the unbonded amount to the StakePool. This also fixes a potential misuse-case, where someone could have donated BNB in excess amount to the protocol, resulting in protocol failure.

### Some other minor code-quality improvements